+++ dispatcher/test/type_table.t.cpp
@@ -8,10 +8,10 @@ using namespace windup;
 
 BOOST_AUTO_TEST_CASE(test_id_to_type)
 {
-  auto msg_a = TypeTable<MsgHolder>::type((int)TypeId::kMsgA);
+  auto msg_a = TypeTable<MsgHolder>::type(TypeId::kMsgA);
   BOOST_REQUIRE(msg_a.which() == 0);
-  auto msg_b = TypeTable<MsgHolder>::type((int)TypeId::kMsgB);
+  auto msg_b = TypeTable<MsgHolder>::type(TypeId::kMsgB);
   BOOST_REQUIRE(msg_b.which() == 1);
-  auto msg_c = TypeTable<MsgHolder>::type((int)TypeId::kMsgC);
+  auto msg_c = TypeTable<MsgHolder>::type(TypeId::kMsgC);
   BOOST_REQUIRE(msg_c.which() == 2);
 }
